requirements.txt
pyyaml==5.3.1
wheel==0.41.3
yamllint==1.32.0
requests==2.31.0
PyGithub==2.1.1
